Move parsing logic to controller
